Eliminated most build errors (gcc,linux,wxGTK).
What is left is mostly:
 - warning: converting double to int
     -> add a cast/to_int or ignore
 - wrong initialization order in ctor
     -> just swap the order to match the class
 - errors about wxCursors
     -> add a function loadResourceCursor


git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@183 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/control/package_list.cpp
@@ -35,12 +35,12 @@ void PackageList::drawItem(DC& dc, int x, int y, size_t item, bool selected) {
 	dc.SetFont(wxFont(12,wxSWISS,wxNORMAL,wxBOLD,false,_("Arial")));
 	dc.GetTextExtent(capitalize(d.package->short_name), &w, &h);
 	pos = align_in_rect(ALIGN_CENTER, RealSize(w,h), rect);
-	dc.DrawText(capitalize(d.package->short_name), pos.x, pos.y + 110);
+	dc.DrawText(capitalize(d.package->short_name), (int)pos.x, (int)pos.y + 110);
 	// draw name
 	dc.SetFont(wxSystemSettings::GetFont(wxSYS_DEFAULT_GUI_FONT));
 	dc.GetTextExtent(d.package->full_name, &w, &h);
 	RealPoint text_pos = align_in_rect(ALIGN_CENTER, RealSize(w,h), rect);
-	dc.DrawText(d.package->full_name, text_pos.x, text_pos.y + 130);
+	dc.DrawText(d.package->full_name, (int)text_pos.x, (int)text_pos.y + 130);
 }
 
 void PackageList::showData(const String& pattern) {
